requirements.txt
fastapi
uvicorn
pygithub
python-dotenv